requirements.txt
requests==2.31.0
schedule==1.2.0
reportlab==4.0.4
python-dotenv==1.0.0
flask==3.0.0
